Start factoring out commonalities and splitting tests
