+++ src/lib.rs
@@ -77,6 +77,3 @@ pub use crate::rpm::*;
 
 #[cfg(test)]
 mod tests;
-
-#[cfg(all(test, feature = "test-with-podman"))]
-mod compat_tests;
